requirements.txt
pandas~=1.5.3
entsoe-py==0.5.9
boto3~=1.26.54
botocore~=1.29.54